requirements.txt
dash==4.0.0
dash-bootstrap-components==2.0.4
pandas==3.0.0
playwright==1.58.0
plotly==6.5.2
PyYAML==6.0.3
python-docx==1.2.0
requests==2.32.5
numpy==2.4.2
python-dateutil==2.9.0.post0
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0
greenlet==3.3.1
idna==3.11
certifi==2026.1.4
charset-normalizer==3.4.4
urllib3==2.6.3
nest-asyncio==1.6.0
packaging==26.0
retrying==1.4.2
typing_extensions==4.15.0
zipp==3.23.0
importlib_metadata==8.7.1
lxml==6.0.2
supabase>=2.0.0
python-dotenv>=1.0.0
